fix: Добавил 1 тест для проверки исключения при обновлении фильма с датой до 28.12.1895. Правки ревьювера: 1.Переименован метод проверяющий пустое поле name у пользователя на setNameIfNotPresent(). 2.Создан отдельный класс NotFoundException для ошибок Пользователь не найден и Ключ не найден. 3.Переместил тесты из FilmorateApplicationTests в FilmControllerTest и UserControllerTest соответственно. SPRINT9 HW
