+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -2,7 +2,7 @@ package ru.yandex.practicum.filmorate.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
-import ru.yandex.practicum.filmorate.exceptions.ValidationException;
+import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
 import ru.yandex.practicum.filmorate.model.User;
 
 import javax.validation.Valid;
@@ -28,9 +28,8 @@ public class UserController {
     @PostMapping
     public User create(@Valid @RequestBody final User user) {
         user.setId(++userId);
-        checkName(user);
+        setNameIfNotPresent(user);
         users.put(user.getId(), user);
-
         log.info("Добавлен новый пользователь: {}.", user.getEmail());
         return user;
     }
@@ -40,15 +39,15 @@ public class UserController {
         User existingUser = users.get(user.getId());
 
         if (existingUser == null) {
-            throw new ValidationException("Пользователь с ID " + user.getId() + " не найден.");
+            throw new NotFoundException("Пользователь с ID " + user.getId() + " не найден.");
         }
-        checkName(user);
+        setNameIfNotPresent(user);
         users.put(user.getId(), user);
         log.info("Обновление пользователя: {}.", existingUser);
         return user;
     }
 
-    private void checkName(User user) {
+    private void setNameIfNotPresent(User user) {
         if (user.getName() == null) {
             user.setName(user.getLogin());
         }
