+++ src/main/java/ru/yandex/practicum/filmorate/exceptions/NotFoundException.java
@@ -1,0 +1,7 @@
+package ru.yandex.practicum.filmorate.exceptions;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
